requirements.txt
# Dependencies
# ------------
# Add here the list of packages that need to be install to run the app.
# This list is to be updated manually as new packages are added to the project.
# Do not use an automated to freeze local dependencies.
# For convenience dev dependencies are in the file: requirements-dev.txt

# Package installer
pip==24.2

# Async packages
httpx==0.27.2
aiodns==3.2.0
aiofiles==24.1.0
aiohttp==3.10.9
async_lru==2.0.4

# Databases: MongoDB
beanie==1.26.0
motor==3.6.0
pymongo==4.9.2

# Databases: Redis
redis==5.1.1

# Web Framework: FastAPI
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
uvicorn==0.31.0
fastapi==0.115.0
jinja2==3.1.4
PyJWT==2.9.0
itsdangerous==2.2.0

# Worker Framework: Celery
celery==5.4.0

# Utils
requests==2.32.3
oauthlib==3.2.2
sentry-sdk[fastapi]==2.15.0
rich==13.9.2
sapx==0.3.0
aioboto3==13.1.1
Pillow==10.4.0

# Third parties
python-facebook-api==0.17.1
